requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0
duckduckgo-search>=7.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
streamlit-autorefresh>=1.0.0
